refactor: resolved git merge changes.
