+++ app/src/main/java/dev/iamwami/app/quotinews/ui/components/NewsImage.kt
@@ -1,8 +1,6 @@
 package dev.iamwami.app.quotinews.ui.components
 
 import android.widget.Toast
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material.Divider
 import androidx.compose.material.Icon
 import androidx.compose.material.IconToggleButton
 import androidx.compose.material.MaterialTheme
@@ -15,7 +13,6 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.unit.dp
 import coil.compose.AsyncImage
 import dev.iamwami.app.quotinews.R
 import dev.iamwami.app.quotinews.model.News
@@ -37,37 +34,5 @@ fun NewsImage(
     )
 }
 
-@Composable
-fun LikeBtn(
-    modifier: Modifier = Modifier,
-    onCheckedBtnState: Boolean,
-    onToggleBtn: (() -> Unit)
-) {
-    val context = LocalContext.current
-    val message = if (onCheckedBtnState) "Added to bookmark" else "removed from bookmark"
-
-    IconToggleButton(
-        checked = onCheckedBtnState,
-        onCheckedChange = {
-            onToggleBtn()
-            Toast.makeText(context, message, Toast.LENGTH_SHORT).show()
-        }
-    ) {
-
-        Icon(
-            imageVector = if (onCheckedBtnState) Icons.Outlined.FavoriteBorder else Icons.Outlined.Favorite,
-            contentDescription = "Heart icon for favourite news"
-        )
-    }
-}
 
 
-/* full width divider with padding*/
-@Composable
-fun PostDivider() {
-    Divider(
-        modifier = Modifier
-            .padding(14.dp),
-        color = MaterialTheme.colors.onSurface.copy(alpha = 0.08f)
-    )
-}
