+++ app/src/main/java/dev/iamwami/app/quotinews/ui/home/HomeRoute.kt
@@ -1,6 +1,8 @@
 package dev.iamwami.app.quotinews.ui.home
 
 import android.content.Context
+import android.os.Build
+import androidx.annotation.RequiresApi
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.material.ScaffoldState
 import androidx.compose.material.rememberScaffoldState
@@ -10,6 +12,7 @@ import androidx.lifecycle.compose.ExperimentalLifecycleComposeApi
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavHostController
 
+@RequiresApi(Build.VERSION_CODES.O)
 @OptIn(ExperimentalLifecycleComposeApi::class)
 @Composable
 fun HomeRoute(
@@ -24,7 +27,6 @@ fun HomeRoute(
 
     HomeFeedScreenWithNewsList(
         newsFeed = newsFeed,
-        onToggleLikeButton = {},
         onSelectNews = {},
         showTopAppBar = true,
         homeLazyListState = rememberLazyListState(),
