+++ app/src/main/java/dev/iamwami/app/quotinews/ui/home/components/RelatedNews.kt
@@ -25,7 +25,7 @@ import dev.iamwami.app.quotinews.model.News
 import dev.iamwami.app.quotinews.ui.components.LikeBtn
 import dev.iamwami.app.quotinews.ui.components.NewsImage
 import dev.iamwami.app.quotinews.ui.components.PostDivider
-import dev.iamwami.app.quotinews.ui.util.SampleNewsApiDataProvider
+import dev.iamwami.app.quotinews.ui.utils.SampleNewsApiDataProvider
 
 @Composable
 fun RelatedNews(
@@ -59,7 +59,7 @@ fun RelatedNews(
                         style = MaterialTheme.typography.body2,
                     )
 
-                    LikeButton(
+                    LikeBtn(
                         onCheckedBtnState = onToggleLikeBtn.value,
                         onToggleBtn = {
                             onToggleLikeBtn.value = onToggleLikeBtn.value != true
